inline the ad txn processing
