+++ backend/api/src/redeem-market-ad-reward.ts
@@ -1,10 +1,11 @@
 import * as admin from 'firebase-admin'
 
-import { runRedeemBoostFeeTxn, runRedeemBoostTxn } from 'shared/run-txn'
 import { createSupabaseDirectClient } from 'shared/supabase/init'
 import { z } from 'zod'
 import { APIError, authEndpoint, validate } from './helpers'
 import { AD_REDEEM_FEE } from 'common/boost'
+import { FieldValue } from 'firebase-admin/firestore'
+import { MarketAdRedeemFeeTxn, MarketAdRedeemTxn } from 'common/txn'
 
 const schema = z.object({
   adId: z.string(),
@@ -40,10 +41,10 @@ export const redeemboost = authEndpoint(async (req, auth) => {
     [adId]
   )
 
-  const reward = parseFloat(data.cost_per_view)
+  const cost = parseFloat(data.cost_per_view)
   const funds = parseFloat(data.funds)
 
-  if (funds < reward) {
+  if (funds < cost) {
     throw new APIError(
       403,
       'Ad for market does not have enough funds to pay out'
@@ -52,7 +53,26 @@ export const redeemboost = authEndpoint(async (req, auth) => {
 
   // create the redeem txn
   const result = await firestore.runTransaction(async (trans) => {
-    const res = await runRedeemBoostTxn(trans, {
+    const pg = createSupabaseDirectClient()
+
+    await pg.none(
+      `update market_ads 
+      set funds = funds - $1
+      where id = $2`,
+      [cost, adId]
+    )
+
+    const txnColl = firestore.collection(`txns/`)
+
+    const reward = cost - AD_REDEEM_FEE
+
+    const toUserDoc = firestore.doc(`users/${auth.uid}`)
+    trans.update(toUserDoc, {
+      balance: FieldValue.increment(reward),
+      totalDeposits: FieldValue.increment(reward),
+    })
+
+    txnColl.add({
       category: 'MARKET_BOOST_REDEEM',
       fromType: 'AD',
       fromId: adId,
@@ -61,8 +81,10 @@ export const redeemboost = authEndpoint(async (req, auth) => {
       amount: reward - AD_REDEEM_FEE,
       token: 'M$',
       description: 'Redeeming market ad',
-    })
-    await runRedeemBoostFeeTxn(trans, {
+      createdTime: Date.now(),
+    } as MarketAdRedeemTxn)
+
+    txnColl.add({
       category: 'MARKET_BOOST_REDEEM_FEE',
       fromType: 'AD',
       fromId: adId,
@@ -71,14 +93,15 @@ export const redeemboost = authEndpoint(async (req, auth) => {
       amount: AD_REDEEM_FEE,
       token: 'M$',
       description: 'Manifold fee for redeeming market ad',
-    })
-    return res
+      createdTime: Date.now(),
+    } as MarketAdRedeemFeeTxn)
+
+    return { status: 'success' }
   })
 
   if (result.status == 'error') {
     throw new APIError(500, 'An unknown error occurred')
   }
 
-  // return the txn
   return result
 })
